requirements.txt
ultralytics>=8.0.0
supervision>=0.3.0
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pandas>=2.0.0
scikit-learn>=1.0.0
deep-sort-realtime>=1.3.0
python-dotenv>=1.0.0
insightface>=0.7.0
faiss-cpu>=1.7.0
matplotlib>=3.7.0
